- Fixed issue with missing static keyword.
- Test framework: save all results before asserting.
- meta.DefineExpression, ArrayBuilder, InterpolatedStringBuilder.
- Changed default format of ToDisplayString.
- Problem with ?. in template compiler
- Imperative advices don't respect TemplateAttribute properties
- Advices don't set back properties of builders (done for Method only)
- Improved generation patterns for async enumerator
- Fixed issue with incorrect ordering of the pipeline by depth

+++ Tests/Caravela.Framework.Tests.InternalPipeline/Tests/Templating/Syntax/Lambdas/CompileTimeSimpleLambda.t.cs
@@ -1,4 +1,4 @@
 int Method(int a, int b)
 {
-    return (int)2;
+    return 2;
 }
